% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_data.R
\docType{data}
\name{counts}
\alias{counts}
\title{countData}
\format{
A data frame with 33610 rows and 4 variables
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(counts)
}
\description{
a count data
}
\details{
Arabidopsis count data

Data from an Ath experiment

something
}
\examples{
data(counts)
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
