% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIris.R
\name{plotIris}
\alias{plotIris}
\title{A plot iris Function}
\usage{
plotIris(column)
}
\arguments{
\item{column}{The column you want to draw, option: c("Sepal.Length", "Sepal.Width","Petal.Length","Petal.Width")}
}
\description{
This function allows you to draw a column of iris dataset
}
\examples{
plotIris("Sepal.Length")
}
\keyword{plotiris}
